requirements.txt

python==3.9
numpy
matplotlib
pandas
scipy
pytorch
torchinfo
monai
segmentation-models-pytorch
hdf5storage
seaborn
ipywidgets
yacs

https://github.com/zudi-lin/pytorch_connectomics

